Misc updates and bugfixes based on feedback.
Fix crash when cleaning up after playlists. This should be looked into at a later time, as the same issue should in theory happen in the other modules too. The theory is that the garbage collector runs __destruct on classes etc after it has been run on the Spotify class doing sp_session_release()

Add getURI() to all objects.